% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/importance_sampling.R
\name{importance_sampling}
\alias{importance_sampling}
\title{Calculate Importance Weights}
\usage{
importance_sampling(
  originalFit,
  oldData,
  newData,
  model_code = originalFit@stanmodel@model_code[1]
)
}
\arguments{
\item{originalFit}{stanfit containing MCMC samples}

\item{oldData}{list() containing data used to fit originalFit}

\item{newData}{list() containing the new input data}

\item{model_code}{string containing Stan model code for originalFit (optional)}
}
\value{
list containing weights, neff, lpOriginal and lpNew
}
\description{
This function calculates importance weights for a previously fitted Stan
model given a new set of input data. Given a stanfit object with MCMC samples
and a new set of input data for the stan model, calculate importance sampling
weights. For example, a time series model predicting into the future can be
updated with new data post-hoc when data becomes available.
}
